+++ CHANGELOG.md
@@ -74,6 +74,8 @@ These changes are available on the `master` branch, but have not yet been releas
   ([#2138](https://github.com/Pycord-Development/pycord/pull/2138))
 - Added `wait_finish` parameter to `VoiceClient.play` for awaiting the end of a play.
   ([#2194](https://github.com/Pycord-Development/pycord/pull/2194))
+- Added `bridge.command()` decorator as a shortcut to `bridge.bridge_command()`.
+  ([#2212](https://github.com/Pycord-Development/pycord/pull/2212))
 
 ### Changed